Added type dependency to the query key
This commit adds a dependency to the `type` parameter in the query key for the `useQuery` hook in the `ViewColumn` component. The new `type` parameter will default to "list" if not provided, and will be used to determine which API route to call. This ensures that the correct data is fetched and rendered based on the selected `type`.