feat: Add Context to manage filter state and UI components
This commit introduces a `FilterStateProvider`, which includes `FilterStateContext` and `FilterStateUpdateContext` to manage the state of the filter and update it when new tags are selected. It also adds a new UI component `FilterUI` to display the tags selected. `Filter` is now renamed to `FilterSideMenu` and moved to its own directory. The `PageComponent` component also wraps its children with `FilterStateProvider`.