Debug
The error you're seeing, `'` can be escaped with `&apos;`, `&lsquo;`, `&#39;`, `&rsquo;`, is related to a rule in `eslint-plugin-react` called `react/no-unescaped-entities`.

The issue is that you're using a string with apostrophes in it, which can cause issues in JSX. To be more specific, the string "... It's already in here!" is causing the error.

To fix this issue, you can replace the apostrophe (`'`) with its HTML entity equivalent (`&apos;`):

```jsx
<span className='text-md w-full flex-grow truncate text-center text-red-600'>
  {isPresent && "... It&apos;s already in here!"}
</span>
```
Using `&apos;` instead of `'` should solve the problem and eliminate the error.